+++ applications/registration-api/src/main/java/no/dcat/model/Catalog.java
@@ -15,28 +15,13 @@ import java.util.Map;
 @Data
 @ToString(includeFieldNames = false)
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class Catalog {
+public class Catalog extends no.dcat.shared.Catalog {
     public static final String ELASTIC_TYPE = "catalog";
-
-    @Id
-    private String id;
-
-    private String uri;
-
-    // dct:title
-    // Norwegian: Tittel
-    private Map<String,String> title = new HashMap<>();
-
-    //dct:description
-    //Norwegian: Beskrivelse
-    private Map<String,String> description = new HashMap<>();
-
-    private Publisher publisher;
 
     private List<Dataset> dataset;
 
     public Catalog(String orgnr) {
-        this.id = orgnr;
+        this.setId(orgnr);
     }
 
     public Catalog() {
