+++ applications/registration-api/src/main/java/no/dcat/rdf/RdfMessageConverter.java
@@ -1,12 +1,12 @@
 package no.dcat.rdf;
 
 import no.dcat.model.Catalog;
+import no.difi.dcat.datastore.domain.dcat.builders.DcatBuilder;
 import org.springframework.http.HttpInputMessage;
 import org.springframework.http.HttpOutputMessage;
 import org.springframework.http.MediaType;
 import org.springframework.http.converter.AbstractHttpMessageConverter;
 import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.http.converter.HttpMessageNotWritableException;
 
 import java.io.IOException;
 import java.nio.charset.Charset;
