+++ libraries/shared/src/main/java/no/dcat/shared/QualityAnnotation.java
@@ -15,11 +15,12 @@ import java.util.Map;
 @ToString(includeFieldNames = false)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class QualityAnnotation {
-    public static final String AccuracyDimension = "Accuracy";
-    public static final String AvailabilityDimension = "Availability";
-    public static final String CompletenessDimension = "Completeness";
-    public static final String CurrentnessDimension = "Currentness";
-    public static final String RelevanceDimension = "Relevance";
+    public static final String isoNS = "http://iso.org/25012/2008/dataquality/";
+    public static final String Accuracy = isoNS + "Accuracy";
+    public static final String Availability = isoNS + "Availability";
+    public static final String Completeness = isoNS + "Completeness";
+    public static final String Currentness = isoNS + "Currentness";
+    public static final String Relevance = isoNS + "Relevance";
 
     private String inDimension;
     private String motivatedBy;
